Hub_Zeus/Manageur/requirements.txt
paho-mqtt==1.5.1
PyYAML
requests
